tests/wpt/tests/css/css-ui/parsing/field-sizing-invalid.html

<!DOCTYPE html>
<title>CSS UI: parsing field-sizing with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-4/#field-sizing">
<link rel="author" title="Kent Tamura" href="mailto:tkent@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_invalid_value('field-sizing', 'none');
test_invalid_value('field-sizing', 'legacy');
test_invalid_value('field-sizing', 'auto');
test_invalid_value('field-sizing', 'normal');
test_invalid_value('field-sizing', 'normal auto');
test_invalid_value('field-sizing', '100%');
test_invalid_value('field-sizing', '10px');
</script>
</body>
